fix : change money type to double precision

+++ docs/migration/balance_history.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS balance_history(
     id_balance_history serial PRIMARY KEY,
-    principal_amount decimal(10, 2) NOT NULL,
-    money_loan decimal(10, 2) NOT NULL,
-    loan_interest decimal(10, 2) NOT NULL
+    principal_amount double precision NOT NULL,
+    money_loan double precision NOT NULL,
+    loan_interest double precision NOT NULL
 );
